refactor(paimon): move MySQL files to mysql subpackage for consistency
- MySqlToPaimonPipeline → mysql/ subpackage
- MySqlToPaimonSqlPipeline → mysql/ subpackage
- MySqlToPaimonPipelineTest → mysql/ subpackage
- Now module structure is: mysql/ and mongo/ subpackages
