+++ flink-paimon-demo/pom.xml
@@ -193,7 +193,7 @@
                             </filters>
                             <transformers>
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>io.sophiadata.flink.paimon.mongo.MongoToPaimonPipeline</mainClass>
+                                    <mainClass>io.sophiadata.flink.paimon.mysql.MySqlToPaimonPipeline</mainClass>
                                 </transformer>
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ServicesResourceTransformer" />
                             </transformers>
